clean up and test end-to-end
