+++ cronq/rabbit_connection.py
@@ -8,10 +8,17 @@ import socket
 import string
 import time
 import urlparse
+import datetime
 
 from cronq.config import RABBITMQ_URL
 
 from haigha.connections.rabbit_connection import RabbitConnection
+from haigha.message import Message
+
+class NullHandler(logging.Handler):
+    def emit(self, record):
+        pass
+
 
 logger = logging.getLogger(__name__)
 
@@ -114,7 +121,11 @@ def connect(**kwargs):
         random_string,
     )
 
-    heartbeat = heartbeat or 1
+    heartbeat = heartbeat or None
+
+    # override connection string sometimes
+    if kwargs.get('heartbeat'):
+        heartbeat = kwargs.get('heartbeat')
 
     close_cb = kwargs.get('close_cb')
 
@@ -134,3 +145,99 @@ def connect(**kwargs):
 
     return conn
 
+
+class CronqConsumer(object):
+    """handles reconnects"""
+
+    def __init__(self):
+        self._connection = None
+        self._channel = None
+
+        # logging
+        self.logger = logging.getLogger(__name__)
+        self.logger.addHandler(NullHandler())
+
+    def _create_connection(self):
+        self._connection = connect(
+            close_cb=self._connection_closed_cb,
+            heartbeat=50000)
+        self._channel = self._connection.channel()
+
+    def connect(self):
+        return self._create_connection()
+
+    @property
+    def connection(self):
+        if self._connection:
+            return self._connection
+
+    @property
+    def channel(self):
+        if self._channel:
+            return self._channel
+
+    def is_connected(self):
+        return self.connection and self.channel
+
+    def make_connected(self, attempts=3):
+        for _ in range(attempts):
+            if self.is_connected():
+                return True
+            else:
+                self.logger.debug("attempt to create connection")
+                self._create_connection()
+                time.sleep(.1)
+        return False
+
+    @staticmethod
+    def tag_from_msg(msg):
+        tag = msg.delivery_info['delivery_tag']
+        return tag
+
+    def ack(self, msg):
+        tag = self.tag_from_msg(msg)
+        if self.is_connected():
+            self.channel.basic.ack(tag)
+
+    def reject(self, msg, requeue=True):
+        tag = self.tag_from_msg(msg)
+        if self.is_connected():
+            self.channel.basic.reject(tag, requeue=requeue)
+
+    def log_message(self, job_id, run_id, message, lvl=logging.INFO):
+        msg = "[cronq_job_id:{0}] [cronq_run_id]:{1} {2}".format(
+            job_id,
+            run_id,
+            message
+        )
+        self.logger.log(lvl, msg)
+
+    def run_something(self, msg):
+        raise NotImplementedError()
+
+    def consume(self, queue):
+        self.make_connected()
+
+        def runner(msg):
+            return self.run_something(msg)
+
+        self.channel.basic.qos(prefetch_count=1)
+        self.channel.basic.consume(
+            queue=queue,
+            consumer=runner,
+            no_ack=False
+        )
+
+        while self.is_connected():
+            self.connection.read_frames()
+
+        return False
+
+    def _connection_closed_cb(self):
+        if self._connection is None:
+            reason = 'unknown'
+        else:
+            reason = self._connection.close_info['reply_text']
+        self.logger.info('Disconnected because: {}'.format(reason))
+        self._connection = None
+        self._channel = None
