+++ cronq/result_aggregator.py
@@ -4,60 +4,29 @@ import logging
 from uuid import UUID
 
 from cronq.backends.mysql import Storage
-from cronq.queue_connection import connect
+from cronq.rabbit_connection import connect, CronqConsumer
 from dateutil.parser import parse
 
 logger = logging.getLogger(__name__)
 
+class CronqAggregator(CronqConsumer):
 
-def setup():
-    conn = connect()
-    if conn is None:
-        return
-    conn._close_cb = create_connection_closed_cb(conn)
+    def __init__(self):
+        super(CronqAggregator, self).__init__()
+        self.storage = Storage()
 
-    # Create message channel
-    channel = conn.channel()
-
-    channel.basic.qos(prefetch_count=1)
-    runner = create_aggregator(channel)
-
-    channel.basic.consume(
-        queue='cronq_results',
-        consumer=runner,
-        no_ack=False,
-    )
-
-    while True:
-        conn.read_frames()
-
-
-def create_connection_closed_cb(connection):
-    def connection_closed_cb():
-        message = "AMQP broker connection closed; close-info: {0}".format(
-            connection.close_info)
-        logger.warning(message)
-    return connection_closed_cb
-
-
-def create_aggregator(channel):
-    storage = Storage()
-
-    def run_something(msg):
+    def run_something(self, msg):
         tag = msg.delivery_info['delivery_tag']
-
-        def ack():
-            channel.basic.ack(tag)
-
         data = json.loads(str(msg.body))
+
+        job_id = data['job_id']
         run_id = UUID(hex=data['run_id'])
 
-        logger.info('[cronq_job_id:{0}] [cronq_run_id:{1}] Write job result to database: {2}'.format(
-            data.get('job_id'), run_id, str(msg.body)
-        ))
+        self.log_message(job_id, run_id,
+            "Write job result to database {}".format(str(msg.body)))
 
-        storage.add_event(
-            data.get('job_id'),
+        self.storage.add_event(
+            job_id,
             parse(data.get('x-send-datetime')),
             run_id.hex,
             data.get('type'),
@@ -65,22 +34,45 @@ def create_aggregator(channel):
             data.get('return_code'),
         )
 
-        logger.info('[cronq_job_id:{0}] [cronq_run_id:{1}] Attempting to update status on job'.format(  # noqa
-            data.get('job_id'), run_id
-        ))
-        storage.update_job_status(
+        self.log_message(job_id, run_id, "Attempting to update status on job")
+
+        self.storage.update_job_status(
             run_id,
             data.get('job_id'),
             parse(data.get('x-send-datetime')),
             data.get('type'),
             data.get('return_code', None))
 
-        logger.info('[cronq_job_id:{0}] [cronq_run_id:{1}] Acking message on job'.format(
-            data.get('job_id'), run_id
-        ))
-        ack()
+        self.log_message(job_id, run_id, "Acking message on job")
+        self.ack(msg)
 
-    return run_something
+
+def setup():
+    runner = CronqAggregator()
+
+    max_failures = 10
+    while max_failures > 0:
+        runner.connect()
+
+        broken = runner.consume()
+
+        max_failures -= 1
+
+
+    runner.logger.warning("Too many errors, exiting")
+    sys.exit(1)
+
+def setup():
+    runner = CronqAggregator()
+
+    max_failures = 10
+    while max_failures > 0:
+        runner.connect()
+        broken = runner.consume(queue="cronq_results")
+        max_failures -= 1
+
+    runner.logger.warning("Too many errors, exiting")
+    sys.exit(1)
 
 
 def main():
